requirements.txt
# Smart Attendance Face Recognition System
# Python Dependencies

opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
